+++ R/config.R
@@ -1,4 +1,69 @@
 
+#' Set or read global options of the gdalcubes package
+#'
+#' Set global package options to change the default behavior of gdalcubes. These include how many threads are used to process data cubes, how created netCDF files are compressed, and whether
+#' or not debug messages should be printed.
+#'
+#' @param ... not used
+#' @param threads number of threads used to process data cubes
+#' @param ncdf_compression_level integer; compression level for created netCDF files, 0=no compression, 1=fast compression, 9=small compression
+#' @param debug logical;  print debug messages
+#' @param cache logical; TRUE if temporary data cubes should be cached to support fast reprocessing of the same cubes
+#' @param swarm character vector with URLs pointing to gdalcubes_server API endpoints, used to evaluate cube chunks (experimental)
+#' @details 
+#' Data cubes can be processed in parallel where one thread processes one chunk at a time. Setting more threads
+#' than the number of chunks of a cube thus has no effect and will not further reduce computation times.
+#' 
+#' Caching has no effect on disk or memory consumption, 
+#' it simply tries to reuse existing temporary files where possible.
+#' For example, changing only parameters to \code{plot} will not require
+#' rerunning the full data cube operation chain.
+#' 
+#' Passing no arguments will return the current options as a list.
+#' @examples 
+#' gdalcubes_options(threads=4) # set the number of threads
+#' gdalcubes_options() # print current options
+#' @export
+gdalcubes_options <- function(..., threads, ncdf_compression_level, debug, cache, swarm) {
+  if (!missing(threads)) {
+    stopifnot(threads >= 1)
+    stopifnot(threads%%1==0)
+    libgdalcubes_set_threads(threads)
+    .pkgenv$threads = threads
+  }
+  if (!missing(ncdf_compression_level)) {
+    stopifnot(ncdf_compression_level %% 1 == 0)
+    stopifnot(ncdf_compression_level >= 0 && ncdf_compression_level <= 9)
+    .pkgenv$compression_level = ncdf_compression_level
+  }
+  if (!missing(debug)) {
+    stopifnot(is.logical(debug))
+    libgdalcubes_debug_output(debug)
+    .pkgenv$debug = debug
+  }
+  if (!missing(cache)) {
+    stopifnot(is.logical(cache))
+    .pkgenv$use_cube_cache = cache
+  }
+  if (!missing(swarm)) {
+    stopifnot(is.character(swarm))
+    # check whether all endpoints are accessible
+    #libgdalcubes_set_swarm(swarm)
+    warning("swarm mode is currently not supported by the R package")
+    .pkgenv$swarm = swarm
+  }
+  if (nargs() == 0) {
+    return(list(
+      threads = .pkgenv$threads,
+      ncdf_compression_level = .pkgenv$compression_level,
+      debug = .pkgenv$debug,
+      cache = .pkgenv$use_cube_cache
+    ))
+  }
+}
+
+
+
 #' Set the number of threads for parallel data cube processing
 #'
 #' Data cubes can be processed in parallel where one thread processes one chunk at a time. Setting more threads
@@ -6,6 +71,7 @@
 #' @param n number of threads
 #' @examples 
 #' gdalcubes_set_threads(1)
+#' @note THIS FUNCTION IS DEPRECATED AND IS REPLACED BY \code{\link{gdalcubes_options}}.
 #' @export
 gdalcubes_set_threads <- function(n=1) {
   stopifnot(n >= 1)
@@ -29,6 +95,7 @@ gdalcubes_version <- function() {
 #' @examples 
 #' gdalcubes_debug_output(TRUE)
 #' gdalcubes_debug_output(FALSE)
+#' @note THIS FUNCTION IS DEPRECATED AND IS REPLACED BY \code{\link{gdalcubes_options}}.
 #' @export
 gdalcubes_debug_output <- function(debug=TRUE) {
    libgdalcubes_debug_output(debug)
@@ -41,6 +108,7 @@ gdalcubes_debug_output <- function(debug=TRUE) {
 #' @examples 
 #' gdalcubes_set_ncdf_compression(9) # maximum compression
 #' gdalcubes_set_ncdf_compression(0) # no compression
+#' @note THIS FUNCTION IS DEPRECATED AND IS REPLACED BY \code{\link{gdalcubes_options}}.
 #' @export
 gdalcubes_set_ncdf_compression <- function(level=2) {
   stopifnot(level %% 1 == 0)
@@ -77,6 +145,7 @@ gdalcubes_gdalversion <- function() {
 #' @param enable logical, TRUE if you want to use the data cube cache
 #' @examples 
 #' gdalcubes_use_cache(FALSE)
+#' @note THIS FUNCTION IS DEPRECATED AND IS REPLACED BY \code{\link{gdalcubes_options}}.
 #' @export
 gdalcubes_use_cache <- function(enable=TRUE) {
   .pkgenv$use_cube_cache = enable